requirements.txt
python-telegram-bot==12.1.0
logzero
tinydb
bs4
six
